Version 9/requirements.txt
Flask>=3.0,<4
gunicorn>=21,<23
openpyxl>=3.1,<4
pandas>=2.0,<3
numpy>=1.24,<3
scipy>=1.10,<2
matplotlib>=3.8,<4
pennylane>=0.40,<0.45
pennylane-lightning>=0.40,<0.45
qiskit>=2,<3
qiskit-ibm-runtime>=0.46,<1
quimb>=1.8,<2
google-cloud-firestore
google-cloud-secret-manager
google-cloud-storage
google-cloud-run
flask-cors
PyYAML
reportlab>=4,<5
